Move some general utility functions into the util module
